Major CURSED compiler fixes: resolve critical compilation issues
- Fixed LLVM basic block terminator issues by restoring builder insertion points during module loading
- Fixed mathz stdlib compilation with proper parameter types (thicc/i64) for variable compatibility
- Fixed string function runtime integration with proper function signatures
- Resolved parser segmentation faults in complex module loading scenarios
- Implemented working mathz.abs_normie and mathz.abs_meal functions with correct absolute value calculations
- Fixed runtime library linking and function signature compatibility
- Improved test suite progression from complete failures to 23 passing tests with 1 output mismatch

This represents a major milestone in CURSED compiler stability and stdlib functionality.

Amp-Thread-ID: https://ampcode.com/threads/T-42f8a5d3-31e8-4b96-9021-7013c1cf870d

+++ src/rt/cursed_rt.h
@@ -21,7 +21,7 @@ double mathz_add(double a, double b);
 double mathz_sub(double a, double b);
 double mathz_mul(double a, double b);
 double mathz_div(double a, double b);
-double mathz_abs_normie(double x);
+int32_t mathz_abs_normie(int32_t x);
 double mathz_max_normie(double a, double b);
 double mathz_min_normie(double a, double b);
